backend_nsfw/requirements.txt
transformers>=4.21.0
torch>=1.12.0
Pillow>=9.0.0
fastapi>=0.85.0
uvicorn[standard]>=0.18.0
python-multipart>=0.0.5